+++ .website/components/spotlight.vue
@@ -1,69 +1,16 @@
 <script setup lang="ts">
 import { defineProps, onMounted } from 'vue'
-import { Spotlight } from './data/spotlights.ts'
-
-const props = defineProps<{
-    spotlights: Spotlight[]
-}>()
 
 const isOdd = props.spotlights.length % 2 !== 0
 console.log(isOdd)
 </script>
 
 <template>
-    <div id="spotlights" class="flex items-center justify-center">
-        <div  class="container grid grid-cols-6 gap-16">
-            <div 
-                v-for="(item, index) in props.spotlights" 
-                :id="'spot-' + index" 
-                :key="item.title" 
-                class="flex col-span-6 gap-4 rounded-xl p-14 justify-between relative overflow-hidden" 
-                :class="[
-                    isOdd && index === props.spotlights.length - 1 ? [
-                        'flex-row',
-                        '',
-                        'h-96',
-                        'min-h-96'
-                    ] : [
-                        'flex-col',
-                        'md:col-span-3',
-                        'min-h-192'
-                    ]
-                ]" 
-                :style='{
-                    backgroundColor: `${item.color}`,
-                    color: `${item.textColor}`
-                }'
-            >
-                <div class="flex flex-col gap-4" :class="isOdd && index === props.spotlights.length - 1 ? 'justify-center' : ''">
-                    <h1 class="text-3xl font-semibold">{{ item.title }}</h1>
-                    <p class="text-lg">{{ item.subtitle }}</p>
-                    <a 
-                        :href="item.href" 
-                        class="cta bg-white w-fit px-6 py-4 rounded-full font-medium hover:bg-transparent transition-all border-2 border-white"
-                        :style='{
-                            color: `${item.color}`
-                        }'
-                    >
-                        {{ item.cta }}
-                    </a>
-                </div>
-                <img 
-                    v-if="item.src"
-                    :src="item.src" 
-                    :alt="item.alt" 
-                    class="rounded-lg" 
-                    :class="isOdd && index === props.spotlights.length - 1 ? 'w-1/2' : 'w-full'"
-                />
-                <img src="/feather.png" alt="feather" class="absolute right-16 w-16" :class="isOdd && index === props.spotlights.length - 1 ? 'top-16' : 'bottom-16'" />
-            </div>
+    <div id="spotlights" class="flex w-full gap-8 flex-col py-16 px-64">
+        <div class="flex justify-center items-center gap-2 border-dashed border-2 border-gray-300 rounded-lg p-4 relative w-full justify-between">
+            <div class="absolute -top-2 left-2 bg-white px-2 text-xs text-gray-400 rounded-md font-mono text-sm tracking-widest">CommunityProvider()</div>
+            <p class="text-base text-pretty text-gray-600">Join our Discord server to get in touch with the community and the team</p>
+            <a href="/discord.html" class="px-8 py-4 hover:shadow-md transition-shadow font-semibold border border-gray-300 rounded-md">Join the community</a>
         </div>
     </div>
 </template>
-
-<style lang="css" scoped>
-
-    .cta:hover {
-        color: white !important;
-    }
-</style>